begin transfering logic into components for better flow
